liblines.hpp: update generate_point() prototype
When generate_point() had a parameter from accepting variables of
type 'int' to of type 'double', the prototype declared in liblines.hpp
needed to be changed. This commit simply updates that prototype with
it's declaration in liblines.cpp.

+++ liblines/liblines.hpp
@@ -25,6 +25,5 @@ using namespace std;
 void coordinate(double xcoor1, double ycoor1, double xcoor2, double ycoor2);
 void slope(double xcoor, double ycoor, double slope);
 void generate_points(double slope, double y_intercept, int max_number, int starting_number);
-void generate_point(double slope, double y_intercept, int number);
+void generate_point(double slope, double y_intercept, double number);
 void write_generated_points_to_file(double slope, double y_intercept, int max_number, int starting_number, string file_name);
-
